+++ src/core/parsers/index.ts
@@ -8,3 +8,5 @@ export { ModuleParser } from './module-parser';
 export { DirectiveParser } from './directive-parser';
 export { PipeParser } from './pipe-parser';
 export { GitRemoteParser } from './git-remote-parser';
+export { TemplateParser } from './template-parser';
+export { StyleParser } from './style-parser';